refactor: split run.rs god-file into modules
