+++ homeboy.json
@@ -570,8 +570,8 @@
         "core_boundary_leak:core-agnostic-source::src/core/extension/fingerprint.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `rust` appears at line <line> in behavioral context `run_fingerprint_script`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/extension/fingerprint.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `wordpress` appears at line <line> in behavioral context `run_fingerprint_script`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/extension/grammar_strings.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `rust` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
-        "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `cargo` appears at line <line> in behavioral context `extract_formatting_findings`::CoreBoundaryLeak",
-        "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `self_check_output_is_harness_failure`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run/formatting.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `cargo` appears at line <line> in behavioral context `extract_formatting_findings`::CoreBoundaryLeak",
+        "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run/formatting.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `self_check_output_is_harness_failure`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/extension/runtime_helper.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/extension/test/drift.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `extract_changes_from_diff`::CoreBoundaryLeak",
         "core_boundary_leak:core-agnostic-source::src/core/extension/test/drift.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `php`::CoreBoundaryLeak",
@@ -781,8 +781,6 @@
         "structural::src/core/extension/bench/run.rs::HighItemCount",
         "structural::src/core/extension/execution.rs::HighItemCount",
         "structural::src/core/extension/grammar.rs::GodFile",
-        "structural::src/core/extension/lint/run.rs::GodFile",
-        "structural::src/core/extension/lint/run.rs::HighItemCount",
         "structural::src/core/extension/trace/report.rs::HighItemCount",
         "structural::src/core/fuzz.rs::GodFile",
         "structural::src/core/fuzz.rs::HighItemCount",
@@ -915,8 +913,8 @@
               "core_boundary_leak:core-agnostic-source::src/core/extension/fingerprint.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `rust` appears at line <line> in behavioral context `run_fingerprint_script`::CoreBoundaryLeak",
               "core_boundary_leak:core-agnostic-source::src/core/extension/fingerprint.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `wordpress` appears at line <line> in behavioral context `run_fingerprint_script`::CoreBoundaryLeak",
               "core_boundary_leak:core-agnostic-source::src/core/extension/grammar_strings.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `rust` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
-              "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `cargo` appears at line <line> in behavioral context `extract_formatting_findings`::CoreBoundaryLeak",
-              "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `self_check_output_is_harness_failure`::CoreBoundaryLeak",
+              "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run/formatting.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `cargo` appears at line <line> in behavioral context `extract_formatting_findings`::CoreBoundaryLeak",
+              "core_boundary_leak:core-agnostic-source::src/core/extension/lint/run/formatting.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `self_check_output_is_harness_failure`::CoreBoundaryLeak",
               "core_boundary_leak:core-agnostic-source::src/core/extension/runtime_helper.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `top-level`::CoreBoundaryLeak",
               "core_boundary_leak:core-agnostic-source::src/core/extension/test/drift.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `extract_changes_from_diff`::CoreBoundaryLeak",
               "core_boundary_leak:core-agnostic-source::src/core/extension/test/drift.rs::Core boundary leak (core-agnostic-source) configured ecosystem term `php` appears at line <line> in behavioral context `php`::CoreBoundaryLeak",
